% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularStatisticsDescriptivesWrapper.R
\name{CircularStatisticsDescriptives}
\alias{CircularStatisticsDescriptives}
\title{Circular Descriptives}
\usage{
CircularStatisticsDescriptives(
  data = NULL,
  version = "1",
  customPeriod = 360,
  distributionPlot = FALSE,
  distributionPlotHistogram = FALSE,
  distributionPlotMeanVector = FALSE,
  distributionPlotPointStack = FALSE,
  meanDirection = TRUE,
  meanLength = FALSE,
  median = FALSE,
  period = "pi2",
  plotHeight = 320,
  plotWidth = 480,
  range = FALSE,
  splitVariable = list(types = list(), value = ""),
  standardDeviation = TRUE,
  variables = list(types = list(), value = list()),
  variance = FALSE
)
}
\description{
Circular Descriptives
}
