% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularStatisticsMultipleSampleTestsWrapper.R
\name{CircularStatisticsMultipleSampleTests}
\alias{CircularStatisticsMultipleSampleTests}
\title{Circular Multiple-Sample Tests}
\usage{
CircularStatisticsMultipleSampleTests(
  data = NULL,
  version = "1",
  customPeriod = 360,
  dependent = list(types = list(), value = ""),
  fixedFactors = list(types = list(), value = list()),
  harrisonKanji = FALSE,
  period = "pi2",
  plotHeight = 320,
  plotWidth = 480,
  watsonWheeler = TRUE,
  watsonWilliams = FALSE
)
}
\description{
Circular Multiple-Sample Tests
}
