% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircularStatisticsOneSampleTestsWrapper.R
\name{CircularStatisticsOneSampleTests}
\alias{CircularStatisticsOneSampleTests}
\title{Circular One-Sample Tests}
\usage{
CircularStatisticsOneSampleTests(
  data = NULL,
  version = "1",
  customPeriod = 360,
  modifiedRayleigh = FALSE,
  period = "pi2",
  plotHeight = 320,
  plotWidth = 480,
  rao = TRUE,
  raoAlpha = "0.01",
  rayleigh = FALSE,
  splitVariable = list(types = list(), value = ""),
  testValue = 180,
  variables = list(types = list(), value = list()),
  vonMisesCheck = FALSE,
  vonMisesCheckAlpha = "0.01"
)
}
\description{
Circular One-Sample Tests
}
